+++ README.md
@@ -47,3 +47,7 @@
 
 - [x] [FilterTest](study%2Fsrc%2Ftest%2Fjava%2Fservlet%2Fcom%2Fexample%2FFilterTest.java)
 - [x] [ServletTest](study%2Fsrc%2Ftest%2Fjava%2Fservlet%2Fcom%2Fexample%2FServletTest.java)
+
+### 미션 
+
+- [x] Legacy MVC와 @MVC 통합
